fix code-based serverless config
Change from typescript, because our project uses ESM and serverless does not support ESM configuration. By
changing to a javascript file instead we maintain ability to configure dynamically, but also can use the '.cts'
extension signalling that the new cjs version of deploy-tools should be used. Note that, for whatever
reason a .cts extension is not recognized by serverless to do the same thing but keep TS.